falta el menu

+++ app/webroot/js/upload.js
@@ -51,22 +51,26 @@ $(document).ready(function() {
 					var nombre = file[(file.length - 1)];
 					var parentId=$('#pictures-uploader').attr("rel");
 					var controller=$('#pictures-uploader').attr("controller");
+					$.post("/"+controller+"/uploadfy_add", {'name' : nombre,'folder' : "uploads",'parent_id' : parentId}, function(id) {
+						if(id){
+							$("#sortable").append('\
+											<li class="image-container ui-state-default" id="'+id+'">\
+											<div class="image"><img alt="" src="/img/uploads/'+nombre+'"></div>\
+											<div class="actions">\
+											<a onclick="return confirm(\'Are you sure you want to delete # 1?\');" href="/admin/'+controller+'/delete/'+id+'">Borrar</a></div>\
+											</li>\
+							');
+						}
+					});
 					$.post("/images/resizeImage", {name:nombre,folder:'pictures'}, function(data){
 			
-					});
-					$.post("/"+controller+"/uploadfy_add", {'name' : nombre,'folder' : "uploads",'parent_id' : parentId}, function(data) {
-						$(".pictures").append('\
-										<div class="image-container">\
-										<div class="image"><img alt="" src="/img/uploads/'+nombre+'"></div>\
-										<div class="actions">\
-										<a onclick="return confirm(\'Are you sure you want to delete # 1?\');" href="/admin/'+controller+'/delete/'+data+'">Borrar</a></div>\
-										</div>\
-										');
 					});
 
 				},
 				'onAllComplete' : function(event, data) {
 					$('#status-message').text(data.filesUploaded+ ' files uploaded, '+ data.errors+ ' errors.');
+					$( "#sortable" ).sortable('refresh');
+					console.log();
 				}
 			});
